+++ metagenomedb/__init__.py
@@ -1,0 +1,1 @@
+"""Scripts for constructing a metagenome database."""
